backend_code/requirements.txt
tornado~=6.1
numpy~=1.22.4
python-dateutil~=2.8.2
pandas~=1.4.2
nsetools~=1.0.11
nsepy~=0.8